requirements.txt
torch
torchvision
opencv-python
numpy
scikit-learn
matplotlib
seaborn
rembg
Pillow
fastapi
uvicorn[standard]
python-multipart
